requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
<<<<<<< HEAD
django-filter-2.4.0
=======
django-filter==2.4.0
>>>>>>> 11b1ac12f7210423c0812e19f110c0702181da88
